update wisp server and docs, don't redownload wisp client

+++ client/tools/wisp_client.sh
@@ -6,11 +6,15 @@ set -e
 
 VERSION="0.3.3"
 PREFIX=$(realpath build/wisp-js)
-TARBALL_PATH="$(realpath build/wisp-archive.tgz)"
+TARBALL_PATH="$(realpath build/wisp-archive.tar.gz)"
 TARBALL_URL=$(npm view @mercuryworkshop/wisp-js@$VERSION dist.tarball)
+
+if [ -d "$PREFIX" ]; then
+  exit
+fi
 
 rm -rf $PREFIX
 mkdir -p $PREFIX
 wget $TARBALL_URL -O $TARBALL_PATH
-tar xvf $TARBALL_PATH -C $PREFIX
+tar xf $TARBALL_PATH -C $PREFIX
 rm $TARBALL_PATH
